requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
python-dotenv>=0.17
gunicorn>=20.1.0
